+++ tests/end2end/features/commandline/test_commandline_bdd.py
@@ -5,12 +5,15 @@
 # License: GNU GPL v3, see the "LICENSE" and "AUTHORS" files for details.
 
 import inspect
+from contextlib import suppress
 
 from PyQt5.QtCore import Qt
 
 import pytest_bdd as bdd
 
+import vimiv
 from vimiv import api
+from vimiv.commands import runners
 from vimiv.gui import statusbar
 
 
@@ -24,6 +27,11 @@ def hit_backspace(commandline, qtbot):
     qtbot.wait(10)
 
 
+@bdd.when("I run help <topic>")
+def run_help(topic):
+    runners.run(f"help {topic}", mode=api.modes.current())
+
+
 @bdd.then(bdd.parsers.parse("the text in the command line should be {text}"))
 def check_commandline_text(commandline, text):
     assert commandline.text() == text
@@ -34,8 +42,20 @@ def check_commandline_closed():
     assert api.modes.current() != api.modes.COMMAND
 
 
-@bdd.then(bdd.parsers.parse("the help for '{commandname}' should be displayed"))
-def check_command_help(commandname):
-    command = api.commands.get(commandname, mode=api.modes.current())
-    docstring = inspect.getdoc(command.func)
-    assert docstring in statusbar.statusbar.message.text()
+@bdd.then(bdd.parsers.parse("the help for '{topic}' should be displayed"))
+@bdd.then("the help for '<topic>' should be displayed")
+def check_help(topic):
+    if topic == "vimiv":
+        assert vimiv.__description__ in statusbar.statusbar.message.text()
+        return
+    topic = topic.lower().lstrip(":")
+    with suppress(api.commands.CommandNotFound):
+        command = api.commands.get(topic, mode=api.modes.current())
+        docstring = inspect.getdoc(command.func)
+        assert docstring in statusbar.statusbar.message.text()
+        return
+    with suppress(KeyError):
+        setting = api.settings.get(topic)
+        assert setting.desc in statusbar.statusbar.message.text()
+        return
+    raise AssertionError(f"Topic '{topic}' not found")